pyproject.toml
[tool.poetry]
name = "fix-your-life-agent"
version = "0.1.0"
description = ""
authors = ["private <oskar.elvkull@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.14"
langgraph = "^0.2.61"
langchain-anthropic = "^0.3.1"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
